feat: timestamp versioning for SW built assets
