+++ frontend/vite.config.js
@@ -2,6 +2,23 @@ import frappeui from 'frappe-ui/vite'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import path from 'path'
+import fs from 'fs/promises'
+
+const buildTimestamp = Date.now().toString()
+
+const emitServiceWorker = () => ({
+	name: 'slides-service-worker',
+	apply: 'build',
+	async writeBundle() {
+		const swSourcePath = path.resolve(__dirname, 'src/service-worker.js')
+		const swOutputPath = path.resolve(__dirname, '../slides/www/service-worker.js')
+		const source = await fs.readFile(swSourcePath, 'utf8')
+		const stamped = source.replace(/__BUILD_TIMESTAMP__/g, buildTimestamp)
+
+		await fs.mkdir(path.dirname(swOutputPath), { recursive: true })
+		await fs.writeFile(swOutputPath, stamped)
+	},
+})
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -20,6 +37,7 @@ export default defineConfig({
 			},
 		}),
 		vue(),
+		emitServiceWorker(),
 	],
 	server: {
 		allowedHosts: true,
